Added authentication, logout, user and users requests
/authenticate
Checks if post parameters username, password and captcha are set,
validates captcha,
checks if username is in database,
checks if user already has a session,
checks if the password matches with the stored hash
and finally sets session.

/logout
Checks if a session is already there and destroys it

/user
Checks if there is a session (user is logged in or not)
and sends user credentials

/users
Checks if there is a session (user is logged in or not)
and sends user credentials of all sessions (all logged
in users)

added dependencies for session management, password
encryption and post request parsing.